Objective Function Implemented

+++ code/psomp/Particle.h
@@ -40,23 +40,23 @@ class Particle {
 		void setPosition(vector_punteros_a_punto newPosition);
 		void setBestPosition(vector_punteros_a_punto newBestPosition);
 		void setVelocity(vector_punteros_a_punto newVelocity);
-		void setPositionFitness(int newFitness);
-		void setBestPositionFitness(int newFitness);
+		void setPositionFitness(float newFitness);
+		void setBestPositionFitness(float newFitness);
 		void setSize(int newSize);
 
 		vector_punteros_a_punto getPosition();
 		vector_punteros_a_punto getBestPosition();
 		vector_punteros_a_punto getVelocity();
-		int getPositionFitness();
-		int getBestPositionFitness();
+		float getPositionFitness();
+		float getBestPositionFitness();
 		int getSize();
 
 	private:
 		vector_punteros_a_punto position; /* vector de posiciones que definen la ruta */
 		vector_punteros_a_punto bestPosition; /* vector que guarda la mejor posicion */
 		vector_punteros_a_punto velocity; /* vector de velocidades */
-		int positionFitness; /* valor de la funcion objetivo actual */
-		int bestPositionFitness; /* el valor de la f.o. de la mejor posicion */
+		float positionFitness; /* valor de la funcion objetivo actual */
+		float bestPositionFitness; /* el valor de la f.o. de la mejor posicion */
 		int size; /* tamano de la ruta */
 };
